+++ sdk-lib/src/androidTest/java/cash/z/ecc/android/sdk/internal/SdkSynchronizerTest.kt
@@ -5,10 +5,9 @@ import androidx.test.platform.app.InstrumentationRegistry
 import cash.z.ecc.android.bip39.Mnemonics
 import cash.z.ecc.android.sdk.Synchronizer
 import cash.z.ecc.android.sdk.WalletInitMode
+import cash.z.ecc.android.sdk.fixture.LightWalletEndpointFixture
 import cash.z.ecc.android.sdk.fixture.WalletFixture
 import cash.z.ecc.android.sdk.model.ZcashNetwork
-import cash.z.ecc.android.sdk.model.defaultForNetwork
-import co.electriccoin.lightwallet.client.model.LightWalletEndpoint
 import kotlinx.coroutines.test.runTest
 import java.util.UUID
 import kotlin.test.Test
@@ -28,7 +27,7 @@ class SdkSynchronizerTest {
                 InstrumentationRegistry.getInstrumentation().context,
                 ZcashNetwork.Mainnet,
                 alias,
-                LightWalletEndpoint.defaultForNetwork(ZcashNetwork.Mainnet),
+                LightWalletEndpointFixture.newEndpointForNetwork(ZcashNetwork.Mainnet),
                 Mnemonics.MnemonicCode(WalletFixture.SEED_PHRASE).toEntropy(),
                 birthday = null,
                 // Using existing wallet init mode as simplification for the test
@@ -39,7 +38,7 @@ class SdkSynchronizerTest {
                         InstrumentationRegistry.getInstrumentation().context,
                         ZcashNetwork.Mainnet,
                         alias,
-                        LightWalletEndpoint.defaultForNetwork(ZcashNetwork.Mainnet),
+                        LightWalletEndpointFixture.newEndpointForNetwork(ZcashNetwork.Mainnet),
                         Mnemonics.MnemonicCode(WalletFixture.SEED_PHRASE).toEntropy(),
                         birthday = null,
                         // Using existing wallet init mode as simplification for the test
@@ -64,7 +63,7 @@ class SdkSynchronizerTest {
                 InstrumentationRegistry.getInstrumentation().context,
                 ZcashNetwork.Mainnet,
                 alias,
-                LightWalletEndpoint.defaultForNetwork(ZcashNetwork.Mainnet),
+                LightWalletEndpointFixture.newEndpointForNetwork(ZcashNetwork.Mainnet),
                 Mnemonics.MnemonicCode(WalletFixture.SEED_PHRASE).toEntropy(),
                 birthday = null,
                 // Using existing wallet init mode as simplification for the test
@@ -76,7 +75,7 @@ class SdkSynchronizerTest {
                 InstrumentationRegistry.getInstrumentation().context,
                 ZcashNetwork.Mainnet,
                 alias,
-                LightWalletEndpoint.defaultForNetwork(ZcashNetwork.Mainnet),
+                LightWalletEndpointFixture.newEndpointForNetwork(ZcashNetwork.Mainnet),
                 Mnemonics.MnemonicCode(WalletFixture.SEED_PHRASE).toEntropy(),
                 birthday = null,
                 // Using existing wallet init mode as simplification for the test
